Refactor authentication state management to include username and role, and update actions accordingly
