+++ mobile-rider/stores/reducers.ts
@@ -1,4 +1,4 @@
-import authReducer from "./auth/auth-slice";
+import authReducer from "./slices/auth/auth-slice";
 import { combineReducers } from "@reduxjs/toolkit";
 
 const rootReducer = combineReducers({
