+++ mobile-rider/types/auth-types.ts
@@ -11,16 +11,14 @@ export interface RegisterRequest {
 
 export interface AuthResponse {
     token: string;
+    role: string;
     refresh_token: string;
-    user: {
-        id: string;
-        username: string;
-        email: string;
-    };
+    username: string;
 }
 
 export interface AuthState {
-    user: AuthResponse["user"] | null;
+    username: string | null;
+    role: string | null;
     token: string | null;
     loading: boolean;
     error: string | null;
